feat(expenses): allow decimal quantities for line items\n\n- Frontend: quantity input supports decimals (step=any), parsed with parseFloat\n- Backend: validation accepts positive numbers (removes integer constraint)\n- DB: make quantity REAL; add migration 002 for existing DBs compatible with D1 remote\n- Scripts: add db:migrate:002 and db:migrate:002:prod\n\nAlso fixes remote migration by removing explicit transactions/pragmas and recreating index.

+++ migrations/001_initial_schema.sql
@@ -40,7 +40,7 @@ CREATE TABLE line_items (
   id TEXT PRIMARY KEY,
   expense_id TEXT NOT NULL,
   description TEXT NOT NULL,
-  quantity INTEGER NOT NULL,
+  quantity REAL NOT NULL,
   price REAL NOT NULL,
   FOREIGN KEY (expense_id) REFERENCES expenses(id) ON DELETE CASCADE
 );
